rename retrieval & consolidate_sub_answers (initial and refinement)
