+++ backend/onyx/agents/agent_search/deep_search_a/initial/generate_initial_answer/nodes/ingest_initial_base_retrieval.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from onyx.agents.agent_search.deep_search_a.initial.retrieval.states import (
+from onyx.agents.agent_search.deep_search_a.initial.retrieve_orig_question_documents.states import (
     BaseRawSearchOutput,
 )
 from onyx.agents.agent_search.deep_search_a.main.operations import logger
