+++ backend/onyx/agents/agent_search/deep_search_a/initial/retrieve_orig_question_documents/nodes/format_raw_search_results.py
@@ -1,4 +1,4 @@
-from onyx.agents.agent_search.deep_search_a.initial.retrieval.states import (
+from onyx.agents.agent_search.deep_search_a.initial.retrieve_orig_question_documents.states import (
     BaseRawSearchOutput,
 )
 from onyx.agents.agent_search.deep_search_a.shared.expanded_retrieval.states import (
