+++ backend/onyx/context/search/pipeline.py
@@ -24,7 +24,9 @@ from onyx.context.search.models import SearchRequest
 from onyx.context.search.postprocessing.postprocessing import cleanup_chunks
 from onyx.context.search.postprocessing.postprocessing import search_postprocessing
 from onyx.context.search.preprocessing.preprocessing import retrieval_preprocessing
-from onyx.context.search.retrieval.search_runner import retrieve_chunks
+from onyx.context.search.retrieve_orig_question_documents.search_runner import (
+    retrieve_chunks,
+)
 from onyx.context.search.utils import inference_section_from_chunks
 from onyx.context.search.utils import relevant_sections_to_indices
 from onyx.db.models import User
