+++ backend/onyx/context/search/preprocessing/preprocessing.py
@@ -20,7 +20,7 @@ from onyx.context.search.models import SearchRequest
 from onyx.context.search.preprocessing.access_filters import (
     build_access_filters_for_user,
 )
-from onyx.context.search.retrieval.search_runner import (
+from onyx.context.search.retrieve_orig_question_documents.search_runner import (
     remove_stop_words_and_punctuation,
 )
 from onyx.db.engine import CURRENT_TENANT_ID_CONTEXTVAR
